Pivot to using two dialogs
This should be a much more straightforward experience and not create accessibility problems with the theme dropdown dynamically changing the options of the version dropdown.
